*** empty log message ***
git-svn-id: file:///home/rhornig/tmp/svn/inet-framework/trunk@702 539eec35-cc16-0410-b457-356aab9780e1

+++ Network/IPv4/RoutingTable.cc
@@ -237,7 +237,7 @@ void RoutingTable::printRoutingTable()
 
 InterfaceEntry *RoutingTable::interfaceById(int id)
 {
-    if (id<0 || id>=interfaces.size())
+    if (id<0 || id>=(int)interfaces.size())
         opp_error("interfaceById(): nonexistent interface %d", id);
     return interfaces[id];
 }
@@ -422,10 +422,10 @@ int RoutingTable::numRoutingEntries()
 
 RoutingEntry *RoutingTable::routingEntry(int k)
 {
-    if (k < routes.size())
+    if (k < (int)routes.size())
         return routes[k];
     k -= routes.size();
-    if (k < multicastRoutes.size())
+    if (k < (int)multicastRoutes.size())
         return multicastRoutes[k];
     return NULL;
 }
